[feature/ASP-16] Refactor to remove all View objects from this application
