+++ src/main/java/co/pragra/attendancesystem/controller/CohortController.java
@@ -1,66 +1,160 @@
 package co.pragra.attendancesystem.controller;
+
+import co.pragra.attendancesystem.dto.ErrorResponse;
 import co.pragra.attendancesystem.entity.Cohort;
 import co.pragra.attendancesystem.repo.CohortRepo;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
 
+import java.util.Date;
+import java.util.List;
+import java.util.Optional;
 
-@Controller
+// LINE BELOW IS FOR UNIT TESTING PURPOSES ONLY.
+@CrossOrigin(origins = "*")
+
+@RestController
+@Slf4j
+@RequestMapping("/api")
 public class CohortController {
-    private CohortRepo cohortRepo;
 
-    public CohortController(CohortRepo cohortRepo) {
-        this.cohortRepo = cohortRepo;
+    @Autowired
+    private CohortRepo repo;
+
+    public CohortController(CohortRepo repo) {
+        this.repo = repo;
     }
 
-    @GetMapping("/cohorts")
-    public String getAllCohort(Model model){
-        model.addAttribute("cohorts",cohortRepo.findAll());
-        return "cohorts";
+    @GetMapping("/cohort")
+    public ResponseEntity<?> getAllCohort(){
+        List<Cohort> findAll = repo.findAll();
+        if (findAll.size() >= 1) {
+            return ResponseEntity.status(200).body(findAll);
+        } else {
+            ErrorResponse errorR = ErrorResponse.builder()
+                    .appId("AMS-C1")
+                    .errorCode(404)
+                    .dateTime(new Date())
+                    .message("No cohort found in database")
+                    .build();
+            return ResponseEntity.status(404).body(errorR);
+        }
     }
 
-    @GetMapping("/cohort/new")
-    public String addNewCohort(Model model){
-        Cohort cohort = new Cohort();
-        model.addAttribute("cohort",cohort);
-        return "cohort_create";
+    @GetMapping("/cohort/{id}")
+    public ResponseEntity<?> findCohortById(@PathVariable long id){
+        Optional<Cohort> cohortOptional = repo.findById(id);
+        if (cohortOptional.isPresent()) {
+            return ResponseEntity.status(200).body(cohortOptional.get());
+        } else {
+            ErrorResponse errorR = ErrorResponse.builder()
+                    .appId("AMS-C2")
+                    .errorCode(404)
+                    .dateTime(new Date())
+                    .message(String.format("Cohort with ID = [%s] not found in database", id))
+                    .build();
+            return ResponseEntity.status(404).body(errorR);
+        }
     }
 
-    //handing request coming in from adding cohort
-    @PostMapping("/cohorts")
-    public String newCohortAdded(@ModelAttribute Cohort cohort, Model model){
-        //updating repo
-        cohortRepo.save(cohort);
-        //return to main page
-        model.addAttribute("cohorts",cohortRepo.findAll());
-        return "redirect:/cohorts";
+    @GetMapping("/cohort/name/{name}")
+    public ResponseEntity<?> findCohortByName(@PathVariable String name){
+        List<Cohort> list = repo.findCohortByName(name);
+        if (list.size() >= 1) {
+            return ResponseEntity.status(200).body(list);
+        } else {
+            ErrorResponse errorR = ErrorResponse.builder()
+                    .appId("AMS-C3")
+                    .errorCode(404)
+                    .dateTime(new Date())
+                    .message(String.format("Cohort with name = [%s] not found in database", name))
+                    .build();
+            return ResponseEntity.status(404).body(errorR);
+        }
     }
 
-    @GetMapping("/cohort/edit/{id}")
-    public String updatingCohort(@PathVariable("id") Long id, Model model){
-
-        model.addAttribute("cohort",cohortRepo.findById(id).get());
-        //return to updating page
-        return "edit_Cohort";
+    @PostMapping("/cohort")
+    public ResponseEntity<?> createCohort(@RequestBody Cohort cohort){
+        try {
+            Cohort save = repo.save(cohort);
+            return ResponseEntity.status(200).body(save);
+        } catch (Exception e) {
+            // LOG exception
+            ErrorResponse errorR = ErrorResponse.builder()
+                    .appId("AMS-C4")
+                    .errorCode(500)
+                    .dateTime(new Date())
+                    .message(String.format("Cohort not created"))
+                    .build();
+            return ResponseEntity.status(500).body(errorR);
+        }
     }
 
-    @PostMapping("/cohort/edit/{id}")
-    public String editCohort(@ModelAttribute Cohort cohort){
-
-        cohortRepo.save(cohort);
-        //return to main page
-        return "redirect:/cohorts";
+    @PutMapping("/cohort")
+    public ResponseEntity<?> updateCohort(@RequestBody Cohort cohort){
+        Optional<Cohort> byId = repo.findById(cohort.getId());
+        if (byId.isPresent()) {
+            try {
+                Cohort save = repo.save(cohort);
+                return ResponseEntity.status(200).body(save);
+            } catch (Exception e) {
+                // LOG exception
+                ErrorResponse errorR = ErrorResponse.builder()
+                        .appId("AMS-C5")
+                        .errorCode(500)
+                        .dateTime(new Date())
+                        .message(String.format("Cohort not updated"))
+                        .build();
+                return ResponseEntity.status(500).body(errorR);
+            }
+        } else {
+            ErrorResponse errorR = ErrorResponse.builder()
+                    .appId("AMS-C5")
+                    .errorCode(404)
+                    .dateTime(new Date())
+                    .message(String.format("Cohort with ID = [%s] not found in database", cohort.getId()))
+                    .build();
+            return ResponseEntity.status(404).body(errorR);
+        }
     }
 
-    @GetMapping("/cohort/delete/{id}")
-    public String deletingCohort(@PathVariable("id") Long id){
-        cohortRepo.deleteById(id);
-        //return to main page
-        return "redirect:/cohorts";
+    @DeleteMapping("/cohort/{id}")
+    public ResponseEntity<?> deleteCohortById(@PathVariable long id){
+        Optional<Cohort> byId = repo.findById(id);
+        if (byId.isPresent()) {
+            try {
+                repo.deleteById(id);
+                if (!repo.findById(id).isPresent()) {
+                    return ResponseEntity.status(200).body(byId.get());
+                } else {
+                    ErrorResponse errorR = ErrorResponse.builder()
+                            .appId("AMS-C6")
+                            .errorCode(500)
+                            .dateTime(new Date())
+                            .message(String.format("Cohort not deleted"))
+                            .build();
+                    return ResponseEntity.status(500).body(errorR);
+                }
+            } catch (Exception e) {
+                // LOG exception
+                ErrorResponse errorR = ErrorResponse.builder()
+                        .appId("AMS-C6")
+                        .errorCode(500)
+                        .dateTime(new Date())
+                        .message(String.format("Cohort not deleted"))
+                        .build();
+                return ResponseEntity.status(500).body(errorR);
+            }
+        } else {
+            ErrorResponse errorR = ErrorResponse.builder()
+                    .appId("AMS-C6")
+                    .errorCode(404)
+                    .dateTime(new Date())
+                    .message(String.format("Cohort with ID = [%s] not found in database", id))
+                    .build();
+            return ResponseEntity.status(404).body(errorR);
+        }
     }
 }
-
