--- src/main/resources/static/css/cohorts.css
@@ -1,6 +1,0 @@
-#addButton{
-    margin: 20px;
-    margin-left: 20px;
-    color: aliceblue;
-    border-radius: 20px;
-}
