--- src/main/resources/templates/cohort_create.html
@@ -1,67 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="https://www.thymeLeaf.org" lang="en">
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-</head>
-<body>
-<div class="container text-center">
-    <div class="row">
-        <div class="col-lg-6 col-md-6 col-sm-6 container justify-content-center card">
-            <h1 class="text-center">Create New Cohort</h1>
-            <div class="card-body">
-                <form th:action="@{/cohorts}" th:object="${cohort}" method="post">
-
-                    <div class="form-group">
-                        <label>Cohort Name</label>
-                        <input
-                                type="text"
-                                name="name"
-                                th:field="*{name}"
-                                class="form-control"
-                                placeholder="Enter Cohort Name"/>
-                    </div>
-                    <div class="form-group">
-                        <label> Course Name</label>
-                        <input
-                                type="text"
-                                name="courseName"
-                                th:field="*{courseName}"
-                                class="form-control"
-                                placeholder="Enter CourseName Name"/>
-                    </div>
-                    <div class="form-group">
-                        <label for="start">Cohort Start Date:</label>
-                        <input
-                                id="start"
-                                type="date"
-                                name="startDate"
-                                th:field="${cohort.startDate}"
-                                class="form-control"
-                                value="2022-06-01"
-                                min="2018-01-01"/>
-                    </div>
-                    <div class="form-group">
-                        <label for="end"> Course End Date: </label>
-                        <input
-                                id="end"
-                                type="date"
-                                name="endDate"
-                                th:field="${cohort.endDate}"
-                                class="form-control"
-                                value="2022-06-01"
-                                min="2018-01-01"/>
-                    </div>
-                    <div class="box-footer">
-                        <button type="submit" class="btn btn-primary">
-                            Submit
-                        </button>
-                    </div>
-                </form>
-            </div>
-        </div>
-    </div>
-</div>
-</body>
-</html>
