--- src/main/resources/templates/cohorts.html
@@ -1,78 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="https://www.thymeLeaf.org">
-<head>
-  <!-- Required meta tags -->
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-
-  <!-- Bootstrap CSS -->
-  <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.3.1/dist/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
-
-  <!--        Add css file here in case you want-->
-  <link rel="stylesheet" th:href="@{css/cohorts.css}"/>
-  <title>Cohort Management System</title>
-</head>
-<body>
-<!--start of navbar-->
-<nav class="navbar navbar-expand-lg navbar-light bg-warning">
-  <a class="navbar-brand" th:href="@{/cohorts}">Cohort Management</a>
-  <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNavAltMarkup" aria-controls="navbarNavAltMarkup" aria-expanded="false" aria-label="Toggle navigation">
-    <span class="navbar-toggler-icon"></span>
-  </button>
-  <div class="collapse navbar-collapse" id="navbarNavAltMarkup">
-    <div class="navbar-nav">
-      <a class="nav-item nav-link active" th:href="@{/students}">Student Management <span class="sr-only">(current)</span></a>
-      <a class="nav-item nav-link" th:href="@{/#}">Session Management</a>
-      <a class="nav-item nav-link" th:href="@{/#}"> Attendance of Session </a>
-
-    </div>
-  </div>
-</nav>
-<!--end of navbar-->
-
-<div class="row">
-  <div class="col-lg-3">
-    <!--        need to handle this method in student controller while adding new student-->
-    <a th:href ="@{/cohort/new}" class="btn btn-primary btn-sm mb-3" id="addButton"> Add Cohort</a>
-  </div>
-</div>
-
-<!--For getting list-->
-
-<table class="table table-bordered table-striped">
-  <thead class="thead-dark">
-  <tr>
-    <th scope="col">Id</th>
-    <th scope="col">Name of Cohort</th>
-    <th scope="col">Course</th>
-    <th scope="col">Start Date </th>
-    <th scope="col">End Date </th>
-    <th scope="col">Action</th>
-  </tr>
-  </thead>
-  <tbody>
-  <tr th:each="cohort: ${cohorts}">
-    <td th:text = "${cohort.id}"/>
-    <td th:text = "${cohort.name}"/>
-    <td th:text = "${cohort.courseName}"/>
-    <td th:text="${#dates.format(cohort.startDate, 'dd-MMM-yyyy')}"/>
-    <td th:text="${#dates.format(cohort.endDate, 'dd-MMM-yyyy')}"/>
-
-<!--        <td th:text = "${cohort.startDate}"/>-->
-<!--        <td th:text = "${cohort.endDate}"/>-->
-
-    <td>
-      <a th:href="@{/cohort/edit/{id}(id=${cohort.id})}" class="btn btn-primary"> Update</a>
-      <a th:href="@{/cohort/delete/{id}(id=${cohort.id})}" class="btn btn-danger"> Delete</a>
-    </td>
-  </tr>
-  </tbody>
-</table>
-
-<!-- Optional JavaScript -->
-<!-- jQuery first, then Popper.js, then Bootstrap JS -->
-<script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
-<script src="https://cdn.jsdelivr.net/npm/popper.js@1.14.7/dist/umd/popper.min.js" integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" crossorigin="anonymous"></script>
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@4.3.1/dist/js/bootstrap.min.js" integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" crossorigin="anonymous"></script>
-</body>
-</html>
