--- src/main/resources/templates/edit_Student.html
@@ -1,45 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="https://www.thymeLeaf.org" lang="en">
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <title>Title</title>
-</head>
-<body>
-<div class="container">
-  <div class="row">
-    <div class="col-lg-6 col-md-6 col-sm-6 container justify-content-center card">
-      <h1 class="text-center">Edit New Student</h1>
-      <div class="card-body">
-        <form th:action="@{/students/edit/{id} (id=${student.id})}" th:object = "${student}" method="POST">
-
-          <div class="form-group">
-            <label>Update First Name</label>
-            <input
-                    type="text"
-                    name="firstName"
-                    th:field="*{firstName}"
-                    class="form-control"
-                    placeholder="Update student First Name"/>
-          </div>
-          <div class="form-group">
-            <label>Update Last Name</label>
-            <input
-                    type="text"
-                    name="lastName"
-                    th:field="*{lastName}"
-                    class="form-control"
-                    placeholder="Update student Last Name"/>
-          </div>
-          <div class="box-footer">
-            <button type="submit" class="btn btn-primary">
-              Submit
-            </button>
-          </div>
-        </form>
-      </div>
-    </div>
-  </div>
-</div>
-</body>
-</html>
