--- src/main/resources/templates/students.html
@@ -1,75 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="https://www.thymeLeaf.org">
-<head>
-  <!-- Required meta tags -->
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-
-  <!-- Bootstrap CSS -->
-  <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.3.1/dist/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
-  <link rel="stylesheet" th:href="@{css/students.css}"/>
-  <title>Attendance Management System</title>
-</head>
-<body>
-<!--start of navbar-->
-<nav class="navbar navbar-expand-lg navbar-light bg-warning">
-  <a class="navbar-brand" th:href="@{/students}">Student Management</a>
-  <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNavAltMarkup" aria-controls="navbarNavAltMarkup" aria-expanded="false" aria-label="Toggle navigation">
-    <span class="navbar-toggler-icon"></span>
-  </button>
-  <div class="collapse navbar-collapse" id="navbarNavAltMarkup">
-
-<!--    mapping direction for the navbarlinks such as Cohort Managemenent,Session Management, Attendance of Session is attached here, need to create it href-->
-<!--    <div class="navbar-nav">-->
-<!--      <a class="nav-item nav-link active" th:href="@{/cohorts}">Cohort Managemenent <span class="sr-only">(current)</span></a>-->
-<!--      <a class="nav-item nav-link" th:href="@{/sessions}">Session Management</a>-->
-<!--      <a class="nav-item nav-link" th:href="@{/attendances}"> Attendance of Session </a>-->
-<!--    </div>-->
-
-    <div class="navbar-nav">
-      <a class="nav-item nav-link active" th:href="@{/cohorts}">Cohort Managemenent <span class="sr-only">(current)</span></a>
-      <a class="nav-item nav-link" th:href="@{/#}">Session Management</a>
-      <a class="nav-item nav-link" th:href="@{/#}"> Attendance of Session </a>
-    </div>
-  </div>
-</nav>
-<!--end of navbar-->
-
-<div class="row">
-  <div class="col-lg-3">
-    <!--        need to handle this method in student controller while adding new student-->
-    <a th:href ="@{/students/new}" class="btn btn-primary btn-sm mb-3" id="addButton"> Add Student</a>
-  </div>
-</div>
-
-<!--For getting list-->
-
-<table class="table table-bordered table-striped">
-  <thead class="thead-dark">
-  <tr>
-    <th scope="col">Id</th>
-    <th scope="col">First Name</th>
-    <th scope="col">Last Name</th>
-    <th scope="col">Action</th>
-  </tr>
-  </thead>
-  <tbody>
-  <tr th:each="student: ${students}">
-    <td th:text = "${student.id}"/>
-    <td th:text = "${student.firstName}"/>
-    <td th:text = "${student.lastName}"/>
-    <td>
-      <a th:href="@{/students/edit/{id}(id=${student.id})}" class="btn btn-primary"> Update</a>
-      <a th:href="@{/students/delete/{id}(id=${student.id})}" class="btn btn-danger"> Delete</a>
-    </td>
-  </tr>
-  </tbody>
-</table>
-
-<!-- Optional JavaScript -->
-<!-- jQuery first, then Popper.js, then Bootstrap JS -->
-<script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
-<script src="https://cdn.jsdelivr.net/npm/popper.js@1.14.7/dist/umd/popper.min.js" integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" crossorigin="anonymous"></script>
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@4.3.1/dist/js/bootstrap.min.js" integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" crossorigin="anonymous"></script>
-</body>
-</html>
